.history/requirements_20250527222211.txt
Flask
flask-cors
pymongo[srv]
python-dotenv
